factors out event tests into new file
