+++ src/event/mod.rs
@@ -1,0 +1,155 @@
+use oscillator::{Oscillator, StereoWaveform};
+use ratios::R;
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct Event {
+    pub frequency: f32,
+    pub ratios: Vec<R>,
+    pub length: f32,
+    pub gain: f32,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct Phrase {
+    pub events: Vec<Event>,
+}
+
+impl Event {
+    pub fn new(frequency: f32, ratios: Vec<R>, length: f32, gain: f32) -> Event {
+        Event {
+            frequency,
+            ratios,
+            length,
+            gain,
+        }
+    }
+}
+
+pub trait Render<T> {
+    fn render(&mut self, oscillator: &mut Oscillator) -> StereoWaveform;
+}
+
+impl Render<Event> for Event {
+    fn render(&mut self, oscillator: &mut Oscillator) -> StereoWaveform {
+        oscillator.update_freq_gain_and_ratios(self.frequency, self.gain, &self.ratios);
+        let n_samples_to_generate = (self.length * 44_100.0).floor() as usize;
+        oscillator.generate(n_samples_to_generate)
+    }
+}
+
+impl Render<Phrase> for Phrase {
+    fn render(&mut self, oscillator: &mut Oscillator) -> StereoWaveform {
+        let mut result: StereoWaveform = StereoWaveform::new(0);
+        for mut event in self.events.clone() {
+            //            println!("{:?}", event);
+            let stereo_waveform = event.render(oscillator);
+            result.append(stereo_waveform);
+        }
+
+        result
+    }
+}
+
+impl Render<Vec<Phrase>> for Vec<Phrase> {
+    fn render(&mut self, oscillator: &mut Oscillator) -> StereoWaveform {
+        let mut result: StereoWaveform = StereoWaveform::new(0);
+        for phrase in self.iter_mut() {
+            let stereo_waveform = phrase.render(oscillator);
+            result.append(stereo_waveform);
+        }
+
+        result
+    }
+}
+
+pub trait Mutate<T> {
+    fn transpose(&mut self, mul: f32, add: f32) -> T;
+    fn mut_ratios(&mut self, ratios: Vec<R>) -> T;
+    fn mut_length(&mut self, mul: f32, add: f32) -> T;
+    fn mut_gain(&mut self, mul: f32, add: f32) -> T;
+}
+
+impl Mutate<Event> for Event {
+    fn transpose(&mut self, mul: f32, add: f32) -> Event {
+        self.frequency = self.frequency * mul + add;
+        self.clone()
+    }
+
+    fn mut_ratios(&mut self, ratios: Vec<R>) -> Event {
+        self.ratios = ratios;
+        self.clone()
+    }
+
+    fn mut_length(&mut self, mul: f32, add: f32) -> Event {
+        self.length = self.length * mul + add;
+        self.clone()
+    }
+
+    fn mut_gain(&mut self, mul: f32, add: f32) -> Event {
+        self.gain = self.gain * mul + add;
+        self.clone()
+    }
+}
+
+impl Mutate<Phrase> for Phrase {
+    fn transpose(&mut self, mul: f32, add: f32) -> Phrase {
+        for event in self.events.iter_mut() {
+            event.transpose(mul, add);
+        }
+        self.clone()
+    }
+
+    fn mut_ratios(&mut self, ratios: Vec<R>) -> Phrase {
+        for event in self.events.iter_mut() {
+            event.mut_ratios(ratios.clone());
+        }
+        self.clone()
+    }
+
+    fn mut_length(&mut self, mul: f32, add: f32) -> Phrase {
+        for event in self.events.iter_mut() {
+            event.mut_length(mul, add);
+        }
+        self.clone()
+    }
+
+    fn mut_gain(&mut self, mul: f32, add: f32) -> Phrase {
+        for event in self.events.iter_mut() {
+            event.mut_gain(mul, add);
+        }
+        self.clone()
+    }
+}
+
+impl Mutate<Vec<Phrase>> for Vec<Phrase> {
+    fn transpose(&mut self, mul: f32, add: f32) -> Vec<Phrase> {
+        for phrase in self.iter_mut() {
+            phrase.transpose(mul, add);
+        }
+        self.clone()
+    }
+
+    fn mut_ratios(&mut self, ratios: Vec<R>) -> Vec<Phrase> {
+        for phrase in self.iter_mut() {
+            phrase.mut_ratios(ratios.clone());
+        }
+        self.clone()
+    }
+
+    fn mut_length(&mut self, mul: f32, add: f32) -> Vec<Phrase> {
+        for phrase in self.iter_mut() {
+            phrase.mut_length(mul, add);
+        }
+        self.clone()
+    }
+
+    fn mut_gain(&mut self, mul: f32, add: f32) -> Vec<Phrase> {
+        for phrase in self.iter_mut() {
+            phrase.mut_gain(mul, add);
+        }
+        self.clone()
+    }
+}
+
+#[cfg(test)]
+mod test;
